# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Puja Booking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/me/devices": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register Device",
        "description": "Upsert by device_token (globally unique). Re-register moves token to current user.",
        "operationId": "register_device_v1_me_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me/devices/{device_token}": {
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Unregister Device",
        "operationId": "unregister_device_v1_me_devices__device_token__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/pujas": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Pujas",
        "description": "Active catalogue with categories, price ranges, and hero images.",
        "operationId": "list_pujas_v1_pujas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPujaListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pujas/{puja_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Puja",
        "description": "Puja detail \u2014 content blocks, addons, and gallery (ready media only).",
        "operationId": "get_puja_v1_pujas__puja_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "puja_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Puja Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPujaDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pujaris": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Available Pujaris",
        "description": "Verified pujaris priced via pricing_resolver for the puja + slot.",
        "operationId": "available_pujaris_v1_pujaris_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "puja_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Puja Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "time",
              "title": "Time"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPujariListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/catalog/pujas": {
      "get": {
        "tags": [
          "admin-catalog"
        ],
        "summary": "List Pujas",
        "operationId": "list_pujas_v1_admin_catalog_pujas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PujaListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-catalog"
        ],
        "summary": "Create Puja",
        "operationId": "create_puja_v1_admin_catalog_pujas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PujaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PujaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/catalog/pujas/{puja_id}": {
      "put": {
        "tags": [
          "admin-catalog"
        ],
        "summary": "Update Puja",
        "operationId": "update_puja_v1_admin_catalog_pujas__puja_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "puja_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Puja Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PujaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PujaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerContentBlock": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "items"
        ],
        "title": "CustomerContentBlock"
      },
...
      "CustomerPujaDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "default_price": {
            "type": "string",
            "title": "Default Price"
          },
          "price_from": {
            "type": "string",
            "title": "Price From"
          },
          "price_to": {
            "type": "string",
            "title": "Price To"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order"
          },
          "hero_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Image Url"
          },
          "content": {
            "items": {
              "$ref": "#/components/schemas/CustomerContentBlock"
            },
            "type": "array",
            "title": "Content"
          },
          "addons": {
            "items": {
              "$ref": "#/components/schemas/CustomerAddon"
            },
            "type": "array",
            "title": "Addons"
          },
          "gallery": {
            "items": {
              "$ref": "#/components/schemas/CustomerGalleryImage"
            },
            "type": "array",
            "title": "Gallery"
          },
          "is_muhurat_bound": {
            "type": "boolean",
            "title": "Is Muhurat Bound",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "slug",
          "default_price",
          "price_from",
          "price_to",
          "display_order"
        ],
        "title": "CustomerPujaDetail"
      },
      "CustomerPujaListResponse": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CustomerCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "pujas": {
            "items": {
              "$ref": "#/components/schemas/CustomerPujaSummary"
            },
            "type": "array",
            "title": "Pujas"
          }
        },
        "type": "object",
        "required": [
          "categories",
          "pujas"
        ],
        "title": "CustomerPujaListResponse"
      },
...
      "CustomerPujariListResponse": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "pujaris": {
            "items": {
              "$ref": "#/components/schemas/CustomerPujari"
            },
            "type": "array",
            "title": "Pujaris"
          }
        },
        "type": "object",
        "required": [
          "pujaris"
        ],
        "title": "CustomerPujariListResponse"
      },
...
      "DeviceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "device_token": {
            "type": "string",
            "title": "Device Token"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_token",
          "platform",
          "last_seen_at"
        ],
        "title": "DeviceOut"
      },
      "DeviceRegister": {
        "properties": {
          "device_token": {
            "type": "string",
            "maxLength": 255,
            "minLength": 10,
            "title": "Device Token"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "android",
                  "ios",
                  "web"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "type": "object",
        "required": [
          "device_token"
        ],
        "title": "DeviceRegister"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PujaCreate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "default_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Default Price"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "name",
          "default_price"
        ],
        "title": "PujaCreate"
      },
...
      "PujaListResponse": {
        "properties": {
          "pujas": {
            "items": {
              "$ref": "#/components/schemas/PujaResponse"
            },
            "type": "array",
            "title": "Pujas"
          }
        },
        "type": "object",
        "required": [
          "pujas"
        ],
        "title": "PujaListResponse"
      },
...
      "PujaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "default_price": {
            "type": "string",
            "title": "Default Price"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "hero_media_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Media Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "slug",
          "default_price",
          "display_order",
          "is_active"
        ],
        "title": "PujaResponse"
      },
      "PujaUpdate": {
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "default_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Price"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "hero_media_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Media Id"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "title": "PujaUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}